SOEDEPT-169 Update jquery ui to be compatible with jquery.
Open Framework manually overrides the jquery version to use
jquery 1.9.1. Doing so makes the jquery ui version that
ships with Drupal incompatible. As a result, Open Framework
also needs to override the jquery ui in order to prevent
various jquery ui related functionalities from breaking.

Includes jquery ui 1.10.3 and overrides jquery ui js and css.
Uses jquery update module as a reference point for how to
do this, but does not follow the same implementation because
it appears that hook_library_alter() cannot be used inside
a theme.

Down the line if this approach proves bug-ridden, it would be
better to either downgrade the jquery version back to 1.8 or
lower to not break compatibility with the jquery ui version
in core OR force Open Framework to depend on jquery_update
module.